first working order
albeit i dont have 500 bucks a month to shell out on unlimited API calls but 100 countries should suffice. perhaps I should do a check where I have a priority A list of countries that deserve API calls and if and when i am creating the geojson objects i check if that i['properties']['ISO_A3'] is in the  list and if it is then call the API_MOD class method get_news and if not then dotn call the API idk just an idea.